pyproject.toml
[project]
name = "s.warden"
version = "0.1.0"
description = "Sistema de remediacion de incidentes con LLM"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "groq>=1.2.0",
    "pydantic-settings",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
